We don't reset the checksum when restoring a DB in prod

+++ restore.py
@@ -26,17 +26,22 @@ if __name__ == '__main__':
 
     import sys
 
-    if len(sys.argv) == 1:
+    FLAG_RESET_VERSION ='--reset-versions'
+
+    reset_versions = FLAG_RESET_VERSION in sys.argv
+    arguments = filter(lambda x : x != FLAG_RESET_VERSION, sys.argv)
+
+    if len(arguments) == 1:
         sys.stdout.write("Env name: ")
         env_name = raw_input()
     else:
-        env_name= sys.argv[1]
+        env_name= arguments[1]
     ENV_DIR = 'instances/'
 
     # We have to load the environment
     environment_dir = os.path.join(ENV_DIR, env_name)
 
-    dumps_to_restore = sys.argv[2:] if len(sys.argv) > 2 else []
+    dumps_to_restore = arguments[2:] if len(arguments) > 2 else []
 
     if not os.path.isdir(environment_dir):
         print "Invalid environment"
@@ -107,10 +112,12 @@ if __name__ == '__main__':
             if filter(lambda x : x, ret.split('\n')):
                 hwid = get_hardware_id()
                 run_script(dbname, "UPDATE sync_server_entity SET hardware_id = '%s'" % hwid)
-                run_script(dbname, "DELETE FROM sync_server_version WHERE sum NOT IN ('88888888888888888888888888888888', '66f490e4359128c556be7ea2d152e03b')")
+                if reset_versions:
+                    run_script(dbname, "DELETE FROM sync_server_version WHERE sum NOT IN ('88888888888888888888888888888888', '66f490e4359128c556be7ea2d152e03b')")
             else:
                 run_script(dbname, "UPDATE sync_client_sync_server_connection SET host = 'localhost', protocol = 'netrpc_gzip', port = %d" % NETRPC_PORT)
-                run_script(dbname, "DELETE FROM sync_client_version WHERE sum NOT IN ('88888888888888888888888888888888', '66f490e4359128c556be7ea2d152e03b')")
+                if reset_versions:
+                    run_script(dbname, "DELETE FROM sync_client_version WHERE sum NOT IN ('88888888888888888888888888888888', '66f490e4359128c556be7ea2d152e03b')")
 
     except (OSError, IOError) as e:
         raise Exception("Unable to access an environment (cause: %s)" % e)
